refactor: add menus schema & migrations + move project spec files
