+++ nest-api/api/src/recipes/test/recipes.controller.spec.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { RecipesController } from './recipes.controller';
+import { RecipesController } from '../recipes.controller';
 
 describe('RecipesController', () => {
   let controller: RecipesController;
